fix(engine): resolve all clippy warnings (clamp patterns, unused imports)
Replace manual clamp patterns with .clamp() calls across cerebellar,
channels, and population modules. Fix unused import (spike_times),
unnecessary casts, manual swap patterns, and other clippy warnings
in analysis and neuron modules.

cargo clippy -- -D warnings now passes cleanly.

+++ engine/src/analysis/surrogates.rs
@@ -147,7 +147,7 @@ pub fn compound_poisson_process(
     for i in 0..n {
         if rng.random::<f64>() < threshold {
             let n_spikes: usize = rng.sample(poisson) as usize;
-            for s in 0..n_spikes as usize {
+            for s in 0..n_spikes {
                 let offset = i + s;
                 if offset < n {
                     train[offset] = 1.0;
